+++ library/src/main/java/dev/arogundade/library/data/models/account/Balance.kt
@@ -2,5 +2,5 @@ package dev.arogundade.library.data.models.account
 
 data class Balance(
     val account: String,
-    val balance: Long
+    val balance: Double
 )
